+++ tests/testthat/test-derive_extreme_event.R
@@ -1,6 +1,6 @@
-# derive_extreme_records ----
+# derive_extreme_event ----
 ## Test 1: `mode` = first ----
-test_that("derive_extreme_records Test 1: `mode` = first", {
+test_that("derive_extreme_event Test 1: `mode` = first", {
   input <- tibble::tribble(
     ~USUBJID, ~PARAMCD,       ~AVALC,        ~ADY,
     "1",      "NO SLEEP",     "N",              1,
@@ -67,7 +67,7 @@ test_that("derive_extreme_records Test 1: `mode` = first", {
 })
 
 ## Test 2: `mode` = last ----
-test_that("derive_extreme_records Test 2: `mode` = last", {
+test_that("derive_extreme_event Test 2: `mode` = last", {
   input <- tibble::tribble(
     ~USUBJID, ~PARAMCD,       ~AVALC,        ~ADY,
     "1",      "NO SLEEP",     "N",              1,
@@ -134,7 +134,7 @@ test_that("derive_extreme_records Test 2: `mode` = last", {
 })
 
 ## Test 3: `source_datasets` works ----
-test_that("derive_extreme_records Test 3: `source_datasets` works", {
+test_that("derive_extreme_event Test 3: `source_datasets` works", {
   adsl <- tibble::tribble(
     ~USUBJID, ~TRTSDTC,
     "1",      "2020-01-01",
@@ -282,7 +282,7 @@ test_that("derive_extreme_records Test 3: `source_datasets` works", {
 })
 
 ## Test 4: event-specific mode ----
-test_that("derive_extreme_records Test 4: event-specific mode", {
+test_that("derive_extreme_event Test 4: event-specific mode", {
   adhy <- tibble::tribble(
     ~USUBJID, ~AVISITN, ~CRIT1FL,
     "1",             1, "Y",
@@ -342,7 +342,7 @@ test_that("derive_extreme_records Test 4: event-specific mode", {
 })
 
 ## Test 5: event_joined() is handled correctly ----
-test_that("derive_extreme_records Test 5: event_joined() is handled correctly", {
+test_that("derive_extreme_event Test 5: event_joined() is handled correctly", {
   adsl <- tibble::tribble(
     ~USUBJID, ~TRTSDTC,
     "1",      "2020-01-01",
@@ -414,7 +414,7 @@ test_that("derive_extreme_records Test 5: event_joined() is handled correctly", 
         event_joined(
           join_vars = exprs(AVALC, ADT),
           join_type = "after",
-          first_cond = AVALC.join == "CR" &
+          first_cond_upper = AVALC.join == "CR" &
             ADT.join >= ADT + 28,
           condition = AVALC == "CR" &
             all(AVALC.join %in% c("CR", "NE")) &
@@ -426,7 +426,7 @@ test_that("derive_extreme_records Test 5: event_joined() is handled correctly", 
         event_joined(
           join_vars = exprs(AVALC, ADT),
           join_type = "after",
-          first_cond = AVALC.join %in% c("CR", "PR") &
+          first_cond_upper = AVALC.join %in% c("CR", "PR") &
             ADT.join >= ADT + 28,
           condition = AVALC == "PR" &
             all(AVALC.join %in% c("CR", "PR", "NE")) &
@@ -517,7 +517,7 @@ test_that("derive_extreme_records Test 5: event_joined() is handled correctly", 
 })
 
 ## Test 6: no tmp_event_nr_var ----
-test_that("derive_extreme_records Test 6: no tmp_event_nr_var", {
+test_that("derive_extreme_event Test 6: no tmp_event_nr_var", {
   adrs <- tibble::tribble(
     ~USUBJID, ~AVISITN, ~AVALC,
     "1",             1, "PR",
@@ -535,14 +535,14 @@ test_that("derive_extreme_records Test 6: no tmp_event_nr_var", {
       event_joined(
         join_vars = exprs(AVALC),
         join_type = "after",
-        first_cond = AVALC.join == "CR",
+        first_cond_upper = AVALC.join == "CR",
         condition = AVALC == "CR",
         set_values_to = exprs(AVALC = "Y")
       ),
       event_joined(
         join_vars = exprs(AVALC),
         join_type = "after",
-        first_cond = AVALC.join %in% c("CR", "PR"),
+        first_cond_upper = AVALC.join %in% c("CR", "PR"),
         condition = AVALC == "PR",
         set_values_to = exprs(AVALC = "Y")
       )
@@ -566,8 +566,9 @@ test_that("derive_extreme_records Test 6: no tmp_event_nr_var", {
     keys = c("USUBJID", "PARAMCD", "AVISITN")
   )
 })
+
 ## Test 7: deprecation of ignore_event_order ----
-test_that("derive_extreme_records Test 7: deprecation of ignore_event_order", {
+test_that("derive_extreme_event Test 7: deprecation of ignore_event_order", {
   adrs <- tibble::tribble(
     ~USUBJID, ~AVISITN, ~AVALC,
     "1",             1, "PR",
@@ -586,14 +587,14 @@ test_that("derive_extreme_records Test 7: deprecation of ignore_event_order", {
         event_joined(
           join_vars = exprs(AVALC),
           join_type = "after",
-          first_cond = AVALC.join == "CR",
+          first_cond_upper = AVALC.join == "CR",
           condition = AVALC == "CR",
           set_values_to = exprs(AVALC = "Y")
         ),
         event_joined(
           join_vars = exprs(AVALC),
           join_type = "after",
-          first_cond = AVALC.join %in% c("CR", "PR"),
+          first_cond_upper = AVALC.join %in% c("CR", "PR"),
           condition = AVALC == "PR",
           set_values_to = exprs(AVALC = "Y")
         )
@@ -605,7 +606,6 @@ test_that("derive_extreme_records Test 7: deprecation of ignore_event_order", {
     ),
     class = "lifecycle_warning_deprecated"
   )
-
   expected <- bind_rows(
     adrs,
     tibble::tribble(
@@ -618,5 +618,43 @@ test_that("derive_extreme_records Test 7: deprecation of ignore_event_order", {
     base = expected,
     compare = actual,
     keys = c("USUBJID", "PARAMCD", "AVISITN")
+  )
+})
+
+# event_joined ----
+## Test 8: deprecation of `first_cond` ----
+test_that("event_joined Test 8: deprecation of `first_cond`", {
+  new_event <- event_joined(
+    join_vars = exprs(AVALC, ADT),
+    join_type = "after",
+    first_cond_upper = AVALC.join == "CR" &
+      ADT.join >= ADT + 28,
+    condition = AVALC == "CR" &
+      all(AVALC.join %in% c("CR", "NE")) &
+      count_vals(var = AVALC.join, val = "NE") <= 1,
+    set_values_to = exprs(
+      AVALC = "CR"
+    )
+  )
+
+  expect_warning(
+    old_event <- event_joined(
+      join_vars = exprs(AVALC, ADT),
+      join_type = "after",
+      first_cond = AVALC.join == "CR" &
+        ADT.join >= ADT + 28,
+      condition = AVALC == "CR" &
+        all(AVALC.join %in% c("CR", "NE")) &
+        count_vals(var = AVALC.join, val = "NE") <= 1,
+      set_values_to = exprs(
+        AVALC = "CR"
+      )
+    ),
+    class = "lifecycle_warning_deprecated"
+  )
+
+  expect_equal(
+    old_event,
+    expected = new_event
   )
 })