+++ tests/testthat/test-filter_joined.R
@@ -17,16 +17,17 @@ data <- tibble::tribble(
 )
 
 # filter_joined ----
-## Test 1: filter without first_cond ----
-test_that("filter_joined Test 1: filter without first_cond", {
+## Test 1: filter without first_cond_upper ----
+test_that("filter_joined Test 1: filter without first_cond_upper", {
   actual <-
     filter_joined(
       data,
+      dataset_add = data,
       by_vars = exprs(USUBJID),
       join_vars = exprs(AVISITN, AVALC),
       join_type = "after",
       order = exprs(AVISITN),
-      filter = AVALC == "PR" & AVALC.join %in% c("CR", "PR") &
+      filter_join = AVALC == "PR" & AVALC.join %in% c("CR", "PR") &
         AVISITN < AVISITN.join
     )
 
@@ -48,13 +49,14 @@ test_that("filter_joined Test 2: filter with first_cond", {
   actual <-
     filter_joined(
       data,
+      dataset_add = data,
       by_vars = exprs(USUBJID),
       join_vars = exprs(AVALC),
       join_type = "after",
-      first_cond = AVALC == "CR" &
+      first_cond_upper = AVALC == "CR" &
         AVALC.join == "CR",
       order = exprs(AVISITN),
-      filter = TRUE
+      filter_join = TRUE
     )
 
   expected <- tibble::tribble(
@@ -74,13 +76,14 @@ test_that("filter_joined Test 3: filter with first_cond and summary function", {
   actual <-
     filter_joined(
       data,
+      dataset_add = data,
       by_vars = exprs(USUBJID),
       join_vars = exprs(AVALC),
       join_type = "after",
-      first_cond = AVALC == "PR" &
+      first_cond_upper = AVALC == "PR" &
         AVALC.join %in% c("CR", "PR"),
       order = exprs(AVISITN),
-      filter = count_vals(AVALC.join, "SD") <= 1
+      filter_join = count_vals(AVALC.join, "SD") <= 1
     )
 
   expected <- tibble::tribble(
@@ -95,7 +98,7 @@ test_that("filter_joined Test 3: filter with first_cond and summary function", {
   )
 })
 
-## Test 4: join_type = "all" ----
+## Test 4: join_type = 'all' ----
 test_that("filter_joined Test 4: join_type = 'all'", {
   adae <- tibble::tribble(
     ~USUBJID, ~ADY, ~ACOVFL, ~ADURN,
@@ -113,11 +116,12 @@ test_that("filter_joined Test 4: join_type = 'all'", {
 
   actual <- filter_joined(
     adae,
+    dataset_add = adae,
     by_vars = exprs(USUBJID),
     join_vars = exprs(ACOVFL, ADY),
     join_type = "all",
     order = exprs(ADY),
-    filter = ADURN > 30 & ACOVFL.join == "Y" & ADY >= ADY.join - 7
+    filter_join = ADURN > 30 & ACOVFL.join == "Y" & ADY >= ADY.join - 7
   )
 
   expected <- tibble::tribble(
@@ -133,9 +137,69 @@ test_that("filter_joined Test 4: join_type = 'all'", {
   )
 })
 
+## Test 5: deprecation of `filter` ----
+test_that("filter_joined Test 5: deprecation of `filter`", {
+  expect_warning(
+    actual <-
+      filter_joined(
+        data,
+        dataset_add = data,
+        by_vars = exprs(USUBJID),
+        join_vars = exprs(AVISITN, AVALC),
+        join_type = "after",
+        order = exprs(AVISITN),
+        filter = AVALC == "PR" & AVALC.join %in% c("CR", "PR") &
+          AVISITN < AVISITN.join
+      ),
+    class = "lifecycle_warning_deprecated"
+  )
+
+  expected <- tibble::tribble(
+    ~USUBJID, ~AVISITN, ~AVALC,
+    "1",      1,        "PR",
+    "4",      1,        "PR"
+  )
+
+  expect_dfs_equal(
+    base = expected,
+    compare = actual,
+    keys = c("USUBJID", "AVISITN")
+  )
+})
+
+## Test 6: deprecation of `first_cond` ----
+test_that("filter_joined Test 6: deprecation of `first_cond`", {
+  expect_warning(
+    actual <-
+      filter_joined(
+        data,
+        dataset_add = data,
+        by_vars = exprs(USUBJID),
+        join_vars = exprs(AVALC),
+        join_type = "after",
+        first_cond = AVALC == "CR" &
+          AVALC.join == "CR",
+        order = exprs(AVISITN),
+        filter_join = TRUE
+      ),
+    class = "lifecycle_warning_deprecated"
+  )
+
+  expected <- tibble::tribble(
+    ~USUBJID, ~AVISITN, ~AVALC,
+    "1",      2,        "CR"
+  )
+
+  expect_dfs_equal(
+    base = expected,
+    compare = actual,
+    keys = c("USUBJID", "AVISITN")
+  )
+})
+
 # min_cond ----
-## Test 1: test it ----
-test_that("min_cond, Test 1: test it", {
+## Test 7: minimum is derived correctly ----
+test_that("min_cond Test 7: minimum is derived correctly", {
   data <- tibble::tribble(
     ~USUBJID, ~AVISITN, ~AVALC,
     "1",      1,        "PR",
@@ -172,8 +236,8 @@ test_that("min_cond, Test 1: test it", {
 })
 
 # max_cond ----
-## Test 1: test it ----
-test_that("max_cond, Test 1: test it", {
+## Test 8: maximum is derived correctly ----
+test_that("max_cond Test 8: maximum is derived correctly", {
   data <- tibble::tribble(
     ~USUBJID, ~AVISITN, ~AVALC,
     "1",      1,        "PR",